feat: add args parser
